Delete pipeline-related notifications upon pipeline deletion (#1533)

+++ streampipes-pipeline-management/src/main/java/org/apache/streampipes/manager/pipeline/PipelineManager.java
@@ -25,6 +25,7 @@ import org.apache.streampipes.model.base.NamedStreamPipesEntity;
 import org.apache.streampipes.model.client.user.Permission;
 import org.apache.streampipes.model.pipeline.Pipeline;
 import org.apache.streampipes.model.pipeline.PipelineOperationStatus;
+import org.apache.streampipes.resource.management.NotificationsResourceManager;
 import org.apache.streampipes.storage.api.IPermissionStorage;
 import org.apache.streampipes.storage.api.IPipelineStorage;
 import org.apache.streampipes.storage.management.StorageDispatcher;
@@ -111,7 +112,11 @@ public class PipelineManager {
    * @param pipelineId of pipeline to be deleted
    */
   public static void deletePipeline(String pipelineId) {
-    getPipelineStorage().deletePipeline(pipelineId);
+    var pipeline = getPipeline(pipelineId);
+    if (Objects.nonNull(pipeline)) {
+      getPipelineStorage().deletePipeline(pipelineId);
+      new NotificationsResourceManager().deleteNotificationsForPipeline(pipeline);
+    }
   }
